Bug #896854 staging: android: ion: fix ion debug log format issue
[root cause  ] incorrect format
[changes     ] fix ion debug log format issue
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:common
[test case]:compile pass
[reviewers   ] Wenan Hu
[change_type ] debug_log
[tag_product ] common

fix ion debug log format issue.

+++ drivers/staging/android/ion/ion_carveout_heap.c
@@ -80,7 +80,7 @@ static int ion_carveout_heap_allocate(struct ion_heap *heap,
 		goto err_free_table;
 	}
 	pr_info("%s: heap id: %u, paddr: 0x%llx, size: %lu\n",
-		__func__, heap->id, paddr, size);
+		__func__, heap->id, (u64)paddr, size);
 
 	sg_set_page(table->sgl, pfn_to_page(PFN_DOWN(paddr)), size, 0);
 	buffer->sg_table = table;
@@ -107,7 +107,7 @@ static void ion_carveout_heap_free(struct ion_buffer *buffer)
 	sg_free_table(table);
 	kfree(table);
 	pr_info("%s: heap id: %u, paddr: 0x%llx, size: %zu\n",
-		__func__, heap->id, paddr, buffer->size);
+		__func__, heap->id, (u64)paddr, buffer->size);
 }
 
 static struct ion_heap_ops carveout_heap_ops = {
